+++ src/popup/popup.css
@@ -2,29 +2,35 @@
 
 #container {
   display: grid;
-  grid-template-columns: auto 14rem;
+  grid-template-columns: auto auto 14rem;
   grid-template-rows: 2rem auto;
   gap: 0.5rem 2rem;
   /* No idea why `calc(100vh - 2rem)` doesn't work */
   height: calc(600px - 2rem);
 }
 
-#explorer {
+#address-books {
   grid-column: 1;
+  grid-row: 2;
+  overflow: scroll;
+}
+
+#tree {
+  grid-column: 2;
   grid-row: 2;
   user-select: none;
   overflow: scroll;
 }
 
-#column-2 {
-  grid-column: 2;
+#column-3 {
+  grid-column: 3;
   grid-row: 1;
   grid-row-end: 3;
   position: relative;
   overflow: scroll;
 }
 
-#column-2:before {
+#column-3:before {
   content: "";
   position: absolute;
   top: 0;
