+++ src/popup/popup.html
@@ -8,14 +8,8 @@
       href="../external/fontawesome/css/fontawesome.min.css"
       rel="stylesheet"
     />
-    <link
-      href="../external/fontawesome/css/solid.min.css"
-      rel="stylesheet"
-    />
-    <link
-      href="../external/fontawesome/css/regular.min.css"
-      rel="stylesheet"
-    />
+    <link href="../external/fontawesome/css/solid.min.css" rel="stylesheet" />
+    <link href="../external/fontawesome/css/regular.min.css" rel="stylesheet" />
     <link rel="stylesheet" type="text/css" href="../styles.css" />
     <link rel="stylesheet" type="text/css" media="screen" href="popup.css" />
     <script src="../external/email-addresses/lib/email-addresses.min.js"></script>
@@ -29,7 +23,7 @@
         <button id="btn-bcc" class="action-button" role="button">BCC</button>
         <button id="btn-to" class="action-button" role="button">To</button>
       </div>
-
+      <ul id="address-books"></ul>
       <nav id="explorer" class="tree-nav"></nav>
       <div id="column-2">
         <h2 id="category-title"></h2>
